server/requirements.txt
flask==3.1.2
flask-cors==6.0.1

gunicorn==23.0.0

psycopg2-binary==2.9.10

python-magic==0.4.27

mutagen==1.47.0
